feat: :zap: Repositories

+++ src/main/java/com/educate/creationofdatabase/domain/model/Arl.java
@@ -1,23 +1,23 @@
 package com.educate.creationofdatabase.domain.model;
 
 public class Arl {
-    private Integer id;
+    private Long id;
     private String name;
 
     public Arl() {
 
     }
 
-    public Arl(Integer id, String name) {
+    public Arl(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
